GREAT progress!
With the 1.12 release I've now got a much easier system to deal
with over all I'd say. Functions are super nice. They will really
change how this all works. I just create the model and call the
function.

So now materials are used to decide what a block is. This works with
most things, but not so much for command blocks. Materials have a
maximum length, and commands shoot over that really fast. So I've
got a hackish way of doing it right now with command replacements.
So there is improvements to be had, but the plugin is hardly a
full fledged plugin just yet, and I've got plenty to learn, so all
in all I'd say this is going well. w00t!